refactor: move categorical functions to operations
